refactor: EquationGen.js -> generate/equation-gen.js
