+++ packages/compile/src/generate/equation-gen.js
